refactor: Remove framework dependencies from the application layer refactor: Refactor backup functionality, naming feat: Extract TrackableFileStream as interface, introduce DataBufferFluxTrackableFileStream; Remove Spring from DownloadService feat: Refactor file replication and Game Provider interfaces feat: Add DB uniqueness constraint for FileCopy file path ci: Remove unnecessary maven-failsafe-plugin configuration from POM

+++ backend/src/main/java/dev/codesoapbox/backity/core/discovery/domain/events/GameContentDiscoveryProgressChangedEvent.java
@@ -10,7 +10,7 @@ public record GameContentDiscoveryProgressChangedEvent(
         @NonNull GameProviderId gameProviderId,
         int percentage,
         @NonNull Duration timeLeft,
-        int gamesDiscovered,
+        long gamesDiscovered,
         int gameFilesDiscovered
 ) implements DomainEvent {
 }
