+++ backend/src/main/java/dev/codesoapbox/backity/core/discovery/infrastructure/adapters/driven/messaging/ws/model/GameContentDiscoveryResultWsDto.java
@@ -11,7 +11,7 @@ public record GameContentDiscoveryResultWsDto(
         @NotNull LocalDateTime stoppedAt,
         @NotNull GameContentDiscoveryOutcomeWsDto discoveryOutcome,
         LocalDateTime lastSuccessfulDiscoveryCompletedAt,
-        @NotNull int gamesDiscovered,
+        @NotNull long gamesDiscovered,
         @NotNull int gameFilesDiscovered
 ) {
 }
